Move Reservation and Reservationist to top-level interfaces
