+++ neuro_san/internals/reservations/accumulating_agent_reservationist.py
@@ -15,8 +15,8 @@ from typing import Union
 
 from asyncio import Event
 
-from neuro_san.internals.interfaces.reservation import Reservation
-from neuro_san.internals.interfaces.reservationist import Reservationist
+from neuro_san.interfaces.reservation import Reservation
+from neuro_san.interfaces.reservationist import Reservationist
 from neuro_san.internals.reservations.agent_reservation import AgentReservation
