+++ neuro_san/internals/reservations/reservation_util.py
@@ -16,8 +16,8 @@ from typing import Tuple
 
 from asyncio import Event
 
-from neuro_san.internals.interfaces.reservation import Reservation
-from neuro_san.internals.interfaces.reservationist import Reservationist
+from neuro_san.interfaces.reservation import Reservation
+from neuro_san.interfaces.reservationist import Reservationist
 
 
 class ReservationUtil:
