+++ neuro_san/service/watcher/temp_networks/s3_reservations_storage.py
@@ -26,8 +26,8 @@ from botocore.client import BaseClient
 from botocore.exceptions import ClientError
 from botocore.exceptions import NoCredentialsError
 
+from neuro_san.interfaces.reservation import Reservation
 from neuro_san.internals.interfaces.reservations_storage import ReservationsStorage
-from neuro_san.internals.interfaces.reservation import Reservation
 from neuro_san.service.watcher.interfaces.startable import Startable
 from neuro_san.service.watcher.temp_networks.reservation_dictionary_converter import ReservationDictionaryConverter
